Move local-databases-ui.ts to /databases/local-databases-ui.ts
