+++ extensions/ql-vscode/src/extension.ts
@@ -44,7 +44,7 @@ import {
   TemplateQueryReferenceProvider,
 } from "./language-support";
 import { DatabaseManager } from "./local-databases";
-import { DatabaseUI } from "./local-databases-ui";
+import { DatabaseUI } from "./databases/local-databases-ui";
 import {
   DEFAULT_DISTRIBUTION_VERSION_RANGE,
   DistributionKind,
