+++ extensions/ql-vscode/test/vscode-tests/no-workspace/databases/local-databases-ui.test.ts
@@ -9,10 +9,10 @@ import {
 } from "fs-extra";
 import { Uri } from "vscode";
 
-import { DatabaseUI } from "../../../src/local-databases-ui";
-import { testDisposeHandler } from "../test-dispose-handler";
-import { createMockApp } from "../../__mocks__/appMock";
-import { QueryLanguage } from "../../../src/common/query-language";
+import { DatabaseUI } from "../../../../src/databases/local-databases-ui";
+import { testDisposeHandler } from "../../test-dispose-handler";
+import { createMockApp } from "../../../__mocks__/appMock";
+import { QueryLanguage } from "../../../../src/common/query-language";
 
 describe("local-databases-ui", () => {
   describe("fixDbUri", () => {
